Add files via upload
Final changes made to the encryption algorithm. Added another fundamental layer of security. prior to each rotation and shuffle, the array of characters itself is shuffled, making theoretical mathematical analysis of rotation way harder (if not impossible). Speeds on encryption are very fast, just a few milliseconds for even a large message.